Tests fix to JSON Schema tests only (#2226)
* tests: bump mocha, separate tests and change workflow to only test schemas in new node.js versions

* chore: run format

* tests: add shacl-ld test to package.json

* bump jsonld package version
