+++ .github/workflows/validation.yml
@@ -9,7 +9,7 @@ jobs:
   test:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-node@v3.3.0
+      - uses: actions/checkout@v7
+      - uses: actions/setup-node@v7
       - run: npm ci
-      - run: npm run test:validation
+      - run: npm run test:schema
