+++ package.json
@@ -7,7 +7,8 @@
     "render": "bash ./render.sh && prettier . --write",
     "assertions": "node assertions.js",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "test:validation": "mocha validation/**/*.test.js",
+    "test:schema": "mocha validation/td-validation-schema.test.js",
+    "test:shacl-ld": "mocha validation/td-validation-shacl-ld.test.js",
     "x-validation": "node validation/td-cross-validation.js",
     "generateTmSchema": "node validation/tmSchemaGenerator.js",
     "prepare": "husky install",
@@ -38,9 +39,9 @@
     "csvtojson": "^2.0.8",
     "debug": "^4.1.1",
     "husky": "^7.0.0",
-    "jsonld": "^8.0.0",
-    "mocha": "^10.0.0",
+    "jsonld": "^9.0.0",
     "npm": "^6.14.11",
+    "mocha": "^11.8.0",
     "prettier": "3.0.0",
     "rdf-ext": "^1.3.5",
     "rdf-utils-dataset": "^2.0.0",
